Fixed TBB warnings on setting GPU queue
Solution is not to create context on devices that are not
of the requested device type or not from the requested backend.

```
Python 3.7.7 (default, Jul 14 2020, 22:02:37)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.17.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: import dpctl

In [2]: with dpctl.device_context('opencl:gpu'):
   ...:     print(1)
      ...:
      1

In [3]: with dpctl.device_context('level0:gpu'):
   ...:     print(1)
      ...:
      1

```