scripts/build_backend.py reorg
The logic to build the backend is moved inside a function, allowing
to pass argument controlling the build.

Currently this is used to specify whether to use L0 support (hardcoded to True),
but opens a door to specify custom compiler to use to build SyclInterface library.

setup.py no longer uses subprocess to invoke build_backed, but uses
importlib to import the function defined there and calls the function with
appropriate arguments.

scipts/build_backend.py retains the function of the script, calling the build_backend
function with default arguments (no L0 support)