Use public api keyword in declaring Cython classes
Use api keyword in declaring API functions.

Use of API defines initializer functions which allows native extensions
that work with dpctl Python types to not link to Python-produced shared
objects on Linux, and allows for pybind11 to portably work with Python
types defined in dpctl.

Example for working with dpctl types is to be added to dpctl/examples