Execute body of host_task if main interpreter is still initialized
This PR applies the solution worked out in a toy POC

Host tasks are executed in a separate thread from
the main thread which submitted the kernel and incremented
reference count on objects we want to keep alive.

When kernel submission occurs near the end of the script,
the CPython interpreter may have begun finalization process
by the time the kernel completes the execution.

`PyThread_Ensure` would crash in that case. So execute the host
task only if `Py_IsInitialized()`.

The only testing of this scenario is to submit the kernel at
the end and not call queue synchronization.