Added dpctl.utils.get_coerced_usm_type
This function deduces the type of USM allocation for the output array
based on USM types of the input arrays.

Changed behavior of get_execution_queue(sequence_of_queues).

Queues are equivalent now when they compare equal, i.e. q1 == q2
is True. This means that they are Python wrappers about copies of
the same underlying SYCL queue.

Since dpctl.SyclQueue() and dpctl.SyclQueue() produce two distinct instances
they no longer are equivalent, and will not be compatible in the compute follows
data paradigm.

arr.to_device can be used to zero-copy reattach the new queue to the data provided
that the original queue and the new queue have the same underlying SYCL context,
which guarantees that SYCL pointer can be dereferenced in the new queue.