Adjusted test for reconstruction of USM host allocation from SUAI
The dpctl.MemoryUSM* Python object representing USM allocation are
reconstructed from an object exposing __sycl_usm_array_interface__
use USM-pointer queries. When SUAI provides sycl::context object,
the device is inferred using `get_pointer_device`. For host
allocations this may not need to return the device in the queue used
to make the allocation. It will return the first device in the context
to which the allocation is bound instead (per SYCL 2020 spec).