Use os.add_dll_directory on Windows
This makes DPCTLSyclInterface.dll library findable with Python 3.10

Python 3.10 modifies the behavior on how it looks for DLLs to make it
more secure. Library provides must add it to DLL search path using
'add_dll_directory' utility for Python's internal dynamic library opener
to be able to find it.

This affects loading of native extensions as well, since if native extension
depends on the DLL which can not be found by Python's library loader,
ModuleNotFoundError is raised