dpt.asarray can not migrate arrays across devices
Within compute follows data programming model, we recommend
explicit migration of inputs to the same allocation queue.

This is to be done using `usm_ndarray.to_device` method, or
with `dpctl.tensor.asarray` operation. This operation until
this commit was not able to migrate data across devices.

This commit adds such an ability by copying via host.

Test was added.