Apply work-around for sycl-device-code-split to apply on Windows
In cmake 3.22 using target_link_options result in option inserted
in LINK_FLAGS which places it after -link option and passed directly
to the underlying linker where it is ignored.

This is fixed in cmake 3.25

Also make sure to apply -fsycl-link-huge-device-code on Linux only,
the option is not supported on Windows.