Adjusting test_windows step to work with intel-opencl-rt=2023.1.0
In CI, shell scripts are executed with elevated priviledges, hence
OpenCL loader ignores variable OCL_ICD_FILENAMES. One must set
a registry value. intel-opencl-rt comes with a script to do that.

It also comes with fixed cl.cfg file, so the step to modify cl.cfg
in situ is no longer needed, and in fact corrupts the file.