Add TwoOffsets_CombinedIndexer, and UnpackedStridedIndexer
The TwoOffsets_CombinedIndexer takes two single offset indexers and
combines them into a TwoOffsets struct.

The UnpackedStridedIndexer is a relative of StridedIndexer, except
shapes and strides are provided as separate pointers. This increases
its size, but may be useful to produce indexer that only computes
the second offset using shape/strides stored in packed format for
two-offset indexers.

These are going to be used in reduction kernels.