Return events for computational tasks, rather than temp-clean up host tasks
This resolves a hang that Anton has uncovered. The gist of the bug is

```
for i in range(op_count):
    v[i] = dpt.empty_like(a[i], order='F', dtype=v_type)

    ht_copy_ev[i], copy_ev = ti._copy_usm_ndarray_into_usm_ndarray(src=a[i], dst=v[i], sycl_queue=a_sycl_queue)

    di._wait(copy_ev)
```

Here is `di` is a Pybind11 extensions, whose function `_wait` was implemented as

```
   m.def("_wait", [](sycl::event e) -> void { e.wait(); }, "");
```

The script would hang on the second iteration of the loop if `di._wait` call.
The hang disappears after the change in this commit. Out attempts to build a C++ reproducer
have not been successful thus far.