Implement QOL properties of usm_ndarray
This PR implements request for quality-of-life properties of usm_ndarray requested
in gh-1192. It adds usm_ndarray._element_offset and usm_ndarray._byte_bounds

The _element_offset gives the offset of the start of array from the beginning of
the start of allocation in elements.

_byte_bounds gives the tuple with addresses of end-points of the array. Not every
memory element in between the end-points may be addressed by indexing.