Fixed test_from_dlpack and test_from_dlpack_strided for USM-host allocations
Since USM-host allocation, unlike USM-device and USM-shared allocations,
is not associated with any particular device, only with the host, it is
unreasonable to expect that sycl::get_pointer_device(ptr, ctx) will return
the same device that is stored in the sycl_queue associated with the
dpctl.tensor.usm_ndarray.

The tests are adjusted to bypass that check for USM-host allocations.