Moved inline functions into separate translation units
Instead of using inline keyword to allow multiple definitions of the same function
in different translation units, introduced elementwise_functions_type_utils.cpp
that defines these functions and a header file to use in other translatioon units.

This should reduce the binary size of the produced object files and simplify the
linker's job reducing the link-time.