Fixed gh-1468
Function _reduce_over_axis promotes input array to requested
result data type and carries out reduction computation in that
data type. This is done in dtype if implementation supports it.

If implementation does not support the requested dtype, we reduce
in the default_dtype, and cast to the request dtype afterwards.