Added meta-template arg to submit_(no_)atomic_reduction functions
The last template parameter is a templated class that takes 5
template parameters. This class, instantiated with types, this class
serves as a KernelName for the submitted functor.

The invocation sites were modified to provide such a class as
reduction_*._krn.

The custom_reduction_*_krn class was removed, in favor of using
custom_reduction_wrapper. The generated kernel name, in case
custom reduction functor is called, is custom_reduction_wrapper<KN>,
where KN would be the kernel name for Functor using built-in
sycl::reduce_over_group function.