Change _Memory object's memory ownership model
The _Memory object acquires additional field void * _opaque_ptr.
This pointer is the pointer to std::shared_ptr<void> with a deleter
class to delete USM allocations. The pointer is opaque to retain C-API
of Py_MemoryObject.

If _opaque_ptr is NULL, refobj field is assumed to be responsible for
USM deallocation.

api Memory_GetOpaquePointer is added to access the opaque pointer.

A test in test_sycl_usm is modified to reflect factual changes in
the behavior of _Memory type.