Change vec_sz and n_vecs settings for contiguous kernels
For short data types, each work-item may need to load several
elements to ensure that it uses all the data from cache-line.
For example, with simd32, we load 4 8-bit types (2 cache lines),
2 16-bit types, 1 32-bit and wider types.

n_vec is set to 1, to avoid cache thrashing due to second iteration
of some work-items beginning to access memory at higher addresses
while some work-items continue working on the lower addresses causing
cache evictions.

The size of the work-groups was increated from 128 to 256, which is
chosen so that all 8 threads of single vector with simd32 are used.