Fix lapse in logic of need_copy determination
Closes gh-1926

One-dimensinal contiguous arrays are both C- and F- contiguous.
The check was written with the assumption that f_contig implies
not c_contig, which is untrue for 1D arrays.

The reproducer provided in gh-1926 now outputs

```
astype time: 9.441375732421875e-05
a is a_ True
asarray time: 6.103515625e-05
a is a__ True
```

Previously, astype took several magnitudes longer due to the copy
operation performed.